[refactor] Use src layout like flake8 source code
